Check number of young nodes in a section
Replaced boolean norejectyoung parameter by a numerical one that indicates the number of allowed young peers in a section.

Setting it to 0 means that no control is done.

Additionally added 1 to the number of trailing zeros so that a relocation is done at each event and not just every other event.